+++ common/src/main/java/net/modgarden/silicate/platform/Side.java
@@ -1,0 +1,6 @@
+package net.modgarden.silicate.platform;
+
+public enum Side {
+	CLIENT,
+	DEDICATED_SERVER
+}